// frontend/src/api/models/LLMOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LLMOut = {
    id: number;
    user_id: number;
    provider: string;
    model_name: string;
    base_url: (string | null);
    created_at: string;
    updated_at: string;
};